FossilOrigin-Name: 6759efcc813e66cf937413b4a162eae6bfed4d4b1322721202296d9c9755711d

+++ src/chat.c
@@ -240,6 +240,11 @@ void chat_webpage(void){
   /* New chat messages get inserted immediately after this element */
   @ <span id='message-inject-point'></span>
   @ </div>
+  @ <div id='chat-zoom' class='hidden chat-view'>
+  @  <div id='chat-zoom-content'></div>
+  @  <div class='button-bar'><button class='action-close'>Close Zoom</button></div>
+  @ </div>
+  @ <span id='chat-zoom-marker' class='hidden'><!-- placeholder marker for zoomed msg --></span>
   fossil_free(zProjectName);
   fossil_free(zInputPlaceholder0);
   builtin_fossil_js_bundle_or("popupwidget", "storage", "fetch",
@@ -293,6 +298,7 @@ static const char zChatSchema1[] =
 ** exist.
 */
 void chat_rebuild_index(int bForce){
+  if( !db_table_exists("repository","chat") ) return;
   if( bForce!=0 ){
     db_multi_exec("DROP TABLE IF EXISTS chatfts1");
   }
@@ -319,7 +325,7 @@ void chat_rebuild_index(int bForce){
 ** them if they do not. Set up TEMP triggers (if needed) to update the
 ** chatfts1 table as the chat table is updated.
 */
-static void chat_create_tables(void){
+void chat_create_tables(void){
   if( !db_table_exists("repository","chat") ){
     db_multi_exec(zChatSchema1/*works-like:""*/);
   }else if( !db_table_has_column("repository","chat","lmtime") ){
@@ -1242,6 +1248,15 @@ void chat_msg_from_event(
 **
 **      Show the default URL used to access the chat server.
 **
+** > fossil chat purge
+**
+**      Remove chat messages that are older than chat-keep-days and
+**      which are not one of the most recent chat-keep-count message.
+**
+** > fossil chat reindex
+**
+**      Rebuild the full-text search index for chat
+**
 ** Additional subcommands may be added in the future.
 */
 void chat_command(void){
@@ -1448,6 +1463,15 @@ void chat_command(void){
   }else if( strcmp(g.argv[2],"url")==0 ){
     /* Show the URL to access chat. */
     fossil_print("%s/chat\n", zUrl);
+  }else if( strcmp(g.argv[2],"purge")==0 ){
+    /* clear out expired chat messages:  chat messages that are older then
+    ** chat-keep-days and that are not one or the most recent chat-keep-count
+    ** messages. */
+    chat_create_tables();
+    chat_purge();
+  }else if( strcmp(g.argv[2],"reindex")==0 ){
+    /* Rebuild the FTS5 index on chat content */
+    chat_rebuild_index(1);
   }else{
     fossil_fatal("no such subcommand \"%s\".  Use --help for help", g.argv[2]);
   }
